refactor: :recycle: Icon 컴포넌트로 변경
